% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoRates.Data.R
\name{clear.data}
\alias{clear.data}
\title{DemoRates clear.data: Prepare valid data for demographic transition rates and frequency estimation}
\usage{
clear.data(
  file = NA,
  csv = FALSE,
  para = NA,
  dobm = FALSE,
  dob = FALSE,
  intw = FALSE,
  eventm = FALSE,
  dup = FALSE
)
}
\arguments{
\item{file}{the input Excel data file with spreadsheets Parameters, Person Record and Name (optional).}

\item{csv}{Either \code{TRUE} or \code{FALSE}. Indicate whether the input Person Record is in csv format.}

\item{para}{If the Person Record data is imported in csv format, a separated Excel file with spreadsheets Parameters and Name (optional) should be provided.}

\item{dobm}{Either \code{TRUE} or \code{FALSE}. Details see in \code{dob}.}

\item{dob}{Either \code{TRUE} or \code{FALSE}.
    \code{dobm} controls the imputation/correction option for missing or invalid birth month, while \code{dob} controls the deletion option for missing or invalid birth year and month.

    If \code{dobm} and \code{dob} are both \code{TRUE}, person records with missing or invalid birth year are deleted. Person records with complete birth year but missing or invalid birth month are imputed with a random month as the birth month.

    If \code{dobm = FALSE} and \code{dob = TRUE}, person records with missing or invalid birth year or month are deleted.

    If \code{dobm = TRUE} and \code{dob = FALSE}, person records with missing or invalid birth year are not changed, and person records with complete birth year but missing or invalid birth month are imputed with a random month as birth month.

    If both \code{dobm} and \code{dob} are \code{FALSE}, no modification is made. By default, \code{dobm} and \code{dob} are \code{FALSE}.}

\item{intw}{Either \code{TRUE} or \code{FALSE}. If \code{TRUE}, person records with missing or invalid interview time are imputed with the earliest interview time. If \code{FALSE}, no modification is made. By default, \code{intw = FALSE}.}

\item{eventm}{Either \code{TRUE} or \code{FALSE}. If \code{TRUE}, person records with complete event year but missing or invalid event month are imputed with a random month as event month. If \code{FALSE}, no modification is made. By default, \code{eventm = FALSE}.}

\item{dup}{Either \code{TRUE} or \code{FALSE}. If \code{TRUE}, duplicated events are deleted. If \code{FALSE}, no modification is made. By default, \code{dup = FALSE}.}
}
\description{
The function of \code{clear.data} deals with user-customized Excel files, namely the spreadsheets for parameters, code names and person record. It has the functions of data validation, logic check, data correction, and data conversion. Users could specify the use of function for different types of events by defining the nFunction parameter in the Parameters spreadsheet.
}
