% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoRates.Rate.utility.new.R
\name{run.rates}
\alias{run.rates}
\title{DemoRates run.rates: Estimate demographic transition rates and frequency.}
\usage{
run.rates(
  file = NA,
  csv = FALSE,
  para = NA,
  plot = FALSE,
  sex = TRUE,
  method = NA,
  mfp = TRUE,
  sep.last = TRUE
)
}
\arguments{
\item{file}{Tthe input Excel clear data file output by clear.data function.}

\item{csv}{Either \code{TRUE} or \code{FALSE}. Indicate whether the input data is in csv format.}

\item{para}{If the Person Record data is imported in csv format, a separated Excel file with spreadsheets Parameters and Name (optional) should be provided.}

\item{plot}{Either \code{TRUE} or \code{FALSE}.
If \code{TRUE}, the graph of the comparison between the direct estimations and the Poisson regression results are generated for users. If \code{FALSE}, no plot file is generated. By default, \code{plot = FALSE}.}

\item{sex}{Only available for covariate method. Either \code{TRUE} or \code{FALSE}.
If \code{TRUE}, gender is considered as a factor in estimation. By default, \code{plot = TRUE}.}

\item{method}{Only available for subset method. Choose the estimation model. Either \code{NA}, "Poisson", "Bayes.norm", or "Bayes.log".
If \code{NA}, only direct calculated results will be output. By default, \code{method = NA}.}

\item{mfp}{Choose whether to use mfp package or DemoRates-defined function selection process to run Poisson estimation.
Either \code{TRUE} or \code{FALSE}. If \code{TRUE}, mfp package is used. By default, \code{mfp = TRUE}.}

\item{sep.last}{When split the time interval, decide whether the reminder time period is together with the last time session or not.
Either \code{TRUE} or \code{FALSE}. If \code{TRUE}, the reminder time is separated. By default, \code{sep.last = TRUE}.}
}
\description{
The function \code{run.rates} deals with the cleaned Excel file with spreadsheets of Parameters, Code Names and Person Record. This function conducts Poisson regression model to estimate the transition rates and frequencies of events.
}
